pyproject.toml
[project]
name = "luucy-wallet-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "openai>=1.95.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
